requirements.txt
langchain
openai
python-dotenv
faiss-cpu
pypdf
unstructured